refactor: status attestation flow
